pull: Support require-deltas=upgrade
Our general story for "production" repositories is that people use static
deltas. This is part of a story of better supporting that. In this model we
don't want clients to fall back to per-object fetch if due to some
config/systems failure deltas don't exist.

This will also make it easier for us to potentially introduce a "deltas only"
repository format in the future.