lib/bloom: Add an internal bloom filter implementation
This will be used in an upcoming commit. It adds a basic bloom filter
implementation, using the SipHash family of hash functions.

The implementation (including its parameter choices and hash functions)
will become a protocol detail in future, so must not be changed so that
its output is bitwise incompatible between OSTree versions.

This commit is unpolished and not ready to be merged. It is intended as
a strawman for feedback.

Unit tests are included.

Signed-off-by: Philip Withnall <withnall@endlessm.com>