Update libglnx, port to new GLnxTmpfile
See the rationale in: https://github.com/GNOME/libglnx/pull/46
Basically we should now start unlinking the tmpfile in the non-`O_TMPFILE`
case in more places in the code.

Parts of the commit path get uglier though, but a refactoring there is
a bit overdue; the code is overly complex.

Update submodule: libglnx