fixes in the find remotes async function
need to add the remote to the repo just in case. usually remotes we
get from avahi are not added, so reading configuration in the pulling
code fails.

also, do not clear the error prematurely - that caused the
find-remotes --pull command to "succeed" while actually not pulling
anything at all.