Revert "Add a notion of "physical" sysroot, use for remote writing"
This reverts commit 1eff3e83436b6129c0dc350dbbda52ba330e3834. There
are a few issues with it.  It's not a critical thing for now, so
let's ugly up the git history and revisit when we have time to
debug it and add more tests.

Besides the below issue, I noticed that the simple `ostree remote add`
now writes to `/ostree/repo/config` because we *aren't* using the
`--sysroot` argument.

Closes: https://github.com/ostreedev/ostree/issues/901